+++ src/main/java/com/devhive03/Model/DAO/PostReports.java
@@ -11,7 +11,7 @@ public class PostReports {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "report_id", nullable = false)
-    private Integer reportID;
+    private Long reportID;
 
     @ManyToOne
     @JoinColumn(name= "report_user_id", nullable = false)
